producer/requirements.txt
html2text >= 2020.1.16
google-cloud-pubsub >= 2.8.0, < 3.0.0
python-dotenv >= 0.19.1, < 1.0.0
